Move mock to test-utils folder
