+++ src/__test__/utils/plotSpecs/generateViolinSpec.test.js
@@ -1,7 +1,7 @@
 import _ from 'lodash';
 
 import { generateData } from '../../../utils/plotSpecs/generateViolinSpec';
-import { mockCellSets1 as cellSets } from '../../cellSets.mock';
+import { mockCellSets1 as cellSets } from '../../test-utils/cellSets.mock';
 
 describe('generateData', () => {
   const MOCK_RANDOM = 0.1;
